pyproject.toml
[project]
name = "isaac"
version = "0.2.0"
description = "ISAAC — Intelligent Self-Assembling Agent Constellation. Thin multi-agent terminal harness."
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "rich>=13.9.0",
    "prompt-toolkit>=3.0.48",
    "mcp>=1.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "duckduckgo-search>=7.0",
    "fastembed>=0.4.0",
    "numpy>=1.26.0",
    "litellm>=1.40.0",
]

[project.optional-dependencies]
gateway = [
    "python-telegram-bot>=21.0",
    "flask>=3.0",
]
sandbox = [
    "e2b-code-interpreter>=1.0.0",
]
computer = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
]
wiki = [
    "trafilatura>=1.12.0",
    "pymupdf>=1.24.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8.0",
]

[project.scripts]
isaac = "isaac.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
